bugs fix

+++ src/stores/LayoutStore/LayoutStore.ts
@@ -4,14 +4,11 @@ import {action, computed, flow, makeObservable, observable} from "mobx";
 import {AppToaster, SuccessToast} from "components/Shared";
 import {LayoutConfig, PresetLayout} from "models";
 import {ApiService} from "services";
-import {AlertStore, AppStore, DialogId} from "stores";
+import {AlertStore, AppStore} from "stores";
 
 const MAX_LAYOUT = 10;
 
 export enum LayoutDialogMode {
-    Hidden,
-    Save,
-    Rename,
     DynamicLayout,
     Layout
 }
@@ -34,35 +31,17 @@ export class LayoutStore {
     @observable currentLayoutName: string;
     @observable private layouts: any;
     @observable supportsServer: boolean;
-    @observable oldLayoutName: string | undefined;
     @observable layoutDialogMode: LayoutDialogMode | undefined;
-
-    @computed get isSave(): boolean {
-        return !this.oldLayoutName;
-    }
 
     private constructor() {
         makeObservable<LayoutStore, "layouts">(this);
         this.dockedLayout = null;
         this.layouts = {};
         this.supportsServer = false;
-        this.oldLayoutName = ""; // for rename
         this.initLayoutsFromPresets();
 
         this.layoutDialogMode = LayoutDialogMode.Layout;
     }
-
-    @action setLayoutDialogMode = (mode: LayoutDialogMode) => {
-        this.layoutDialogMode = mode;
-    };
-
-    @action showLayoutDialog = (mode: LayoutDialogMode, oldLayoutName?: string) => {
-        this.setLayoutDialogMode(mode);
-        if (mode === LayoutDialogMode.Rename) {
-            this.setOldLayoutName(oldLayoutName);
-        }
-        AppStore.Instance.dialogStore.showDialog(DialogId.Layout);
-    };
 
     public layoutExists = (layoutName: string): boolean => {
         return layoutName.length > 0 && this.allLayoutNames.includes(layoutName);
@@ -70,10 +49,6 @@ export class LayoutStore {
 
     public setLayoutToBeSaved = (layoutName: string) => {
         this.layoutNameToBeSaved = layoutName ? layoutName : "Empty";
-    };
-
-    public setOldLayoutName = (oldLayoutName: string | undefined) => {
-        this.oldLayoutName = oldLayoutName;
     };
 
     @flow.bound *fetchLayouts() {
@@ -246,8 +221,6 @@ export class LayoutStore {
             return;
         }
 
-        appStore.dialogStore.hideDialog(DialogId.Layout);
-
         // save layout to layouts[] & server/local storage
         const configToSave = this.layouts[oldName];
         this.layouts[newName] = configToSave;
@@ -290,16 +263,19 @@ export class LayoutStore {
             return;
         }
 
-        try {
-            const success = yield appStore.apiService.clearLayout(layoutName);
-            yield dynamicLayout.deleteLayoutMappingByLayoutName(layoutName);
-            if (success) {
-                delete this.layouts[layoutName];
+        const confirmed = yield appStore.alertStore.showInteractiveAlert(`Do you delete layout ${layoutName}?`);
+        if (confirmed) {
+            try {
+                const success = yield appStore.apiService.clearLayout(layoutName);
+                yield dynamicLayout.deleteLayoutMappingByLayoutName(layoutName);
+                if (success) {
+                    delete this.layouts[layoutName];
+                }
+                this.handleDeleteResult(layoutName, success);
+            } catch (err) {
+                console.log(err);
+                this.handleDeleteResult(layoutName, false);
             }
-            this.handleDeleteResult(layoutName, success);
-        } catch (err) {
-            console.log(err);
-            this.handleDeleteResult(layoutName, false);
         }
     }
